setup.py
method 1:
bug check:
possible counter measure against code:
putting bogus "<form" by using comment in source
multiple "<form" will cause code to phreak/break.
solution to counter measure:
1) use reGex module
2) MechanicalSoup / BeautifulSoup

another solution without using modules.
sanitary check by finding "<input" in the form list, we need to do this because all we need is to find the payload so we can test vulnerability

possible countermeasure on "No Module" solution:
commented / bogus "input" element in the form list.
This may cause problem when we start a request with a payload that's not used because it's bogus and it has no use
whether or not requests will still work with a bogus payload is up to me since i haven't tested it, i'm simply thinking

#########################################################################################

removed comments, easy form detect. very line-short.
In conclusion, it's flexible but NOT that flexible.
used reGex module.

method 1 is way flexible but very line-long.